Se mueve la carpeta js y se agrega la clase de usuarios para leer y guardar en el localstorage
