+++ html/signup.html
@@ -91,7 +91,7 @@
         <p class="derechos">Proyecto Integrador - Cesde 2025</p>
     </footer>
 
-    <script src="js/signup.js"></script>
+    <script type="module" src="../js/signup.js"></script>
 </body>
 
 </html>
